+++ akka-stream/src/main/resources/reference.conf
@@ -141,8 +141,8 @@ akka {
     }
 
     serialization-bindings {
-      "akka.stream.scaladsl.SinkRef"    = akka-stream-refs
-      "akka.stream.scaladsl.SourceRef"  = akka-stream-refs
+      "akka.stream.SinkRef"    = akka-stream-refs
+      "akka.stream.SourceRef"  = akka-stream-refs
       "akka.stream.StreamRefs$Protocol" = akka-stream-refs
     }
